Function bp_parachains::parachain_head_storage_key_at_source

pub fn parachain_head_storage_key_at_source(
    paras_pallet_name: &str,
    para_id: ParaId,
) -> StorageKey

Returns runtime storage key of given parachain head at the source chain.

The head is stored by the paras pallet in the Heads map.