Trait bp_parachains::ParaStoredHeaderDataBuilder

pub trait ParaStoredHeaderDataBuilder {
    // Required methods
    fn supported_parachains() -> u32;
    fn try_build(
        para_id: ParaId,
        para_head: &ParaHead
    ) -> Option<ParaStoredHeaderData>;
}

Stored parachain head data builder.

Required Methods

fn supported_parachains() -> u32

Return number of parachains that are supported by this builder.

fn try_build( para_id: ParaId, para_head: &ParaHead ) -> Option<ParaStoredHeaderData>

Try to build head data from encoded head of parachain with given id.