Crate bp_parachains

Primitives of parachains module.

Structs

BestParaHeadHash

Best known parachain head hash.

ImportedParaHeadsKeyProvider

Can be use to access the runtime storage key of the parachain head at the target chain.

ParaInfo

Best known parachain head as it is stored in the runtime storage.

ParaStoredHeaderData

Stored data of the parachain head. It is encoded version of the bp_runtime::StoredHeaderData structure.

ParasInfoKeyProvider

Can be use to access the runtime storage key of the parachains info at the target chain.

SingleParaStoredHeaderDataBuilder

Helper for using single parachain as ParaStoredHeaderDataBuilder.

StoredHeaderData

Header data that we’re storing on-chain.

SubmitParachainHeadsInfo

Info about a SubmitParachainHeads call which tries to update a single parachain.

Enums

BridgeParachainCall

A minimized version of pallet-bridge-parachains::Call that can be used without a runtime.

Traits

ParaStoredHeaderDataBuilder

Stored parachain head data builder.

Functions

parachain_head_storage_key_at_source

Returns runtime storage key of given parachain head at the source chain.

Type Aliases

RelayBlockHash

Block hash of the bridged relay chain.

RelayBlockHasher

Hasher of the bridged relay chain.

RelayBlockNumber

Block number of the bridged relay chain.