Trait bp_parachains::ParaStoredHeaderDataBuilder

pub trait ParaStoredHeaderDataBuilder {
    // Required methods
    fn max_free_head_size() -> u32;
    fn supported_parachains() -> u32;
    fn try_build(
        para_id: ParaId,
        para_head: &ParaHead,
    ) -> Option<ParaStoredHeaderData>;
}

Stored parachain head data builder.

Required Methods

fn max_free_head_size() -> u32

Maximal parachain head size that we may accept for free. All heads above this limit are submitted for a regular fee.

fn supported_parachains() -> u32

Return number of parachains that are supported by this builder.

fn try_build( para_id: ParaId, para_head: &ParaHead, ) -> Option<ParaStoredHeaderData>

Try to build head data from encoded head of parachain with given id.

Object Safety

This trait is not object safe.